requirements.txt
click
PyYAML
Jinja2
